:truck: move download_thread in youwol/environment
